requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.2
PyYAML==6.0.2
aiosqlite==0.20.0
